Change the index buffers to short instead of int

+++ A09 Ex06 Koby 021766944 Inbar 015267479/Dreidel/ObjectModel/CubePosition.cs
@@ -222,7 +222,7 @@ namespace DreidelGame.ObjectModel
         /// </summary>
         private void createIndices()
         {
-            int[] indices = new int[36];  
+            short[] indices = new short[36];  
 
             // Front face
             indices[0] = 0;
@@ -292,11 +292,11 @@ namespace DreidelGame.ObjectModel
 
             iBuffer = new IndexBuffer(
                 this.GraphicsDevice,
-                typeof(int),
+                typeof(short),
                 this.BufferIndices.Length,
                 BufferUsage.WriteOnly);
 
-            iBuffer.SetData<int>(this.BufferIndices);
+            iBuffer.SetData<short>(this.BufferIndices);
 
             this.ComponentVertexBuffer = vBuffer;
             this.ComponentIndexBuffer = iBuffer;
